422: Refactoring Tpp - applying law of dimeter
Tpp used to return it's OIDCRegistrationResponse and allow clients to
then query that object for information. This breaks the law of dimiter.
Instead the Tpp can return data held within it's agregated classes to
reduce coupling. In this case Tpp can now be asked for the Tpp's
registration access token. This means the caller of the Tpp class
doesn't need to know anything about the OIDCRegistration response.

This helps write tests that were required during work to improve
logging and error handling for the DELETE /register/{{ClientId}}
endpoints.

Issue: https://github.com/OpenBankingToolkit/openbanking-aspsp/issues/422