793: Fix more issues with DirectorySoftwareStatement serialisation
I found that when data was deserialised from mongo the iss field was
invisible. This was recreated in the DirectorySoftwareStatementTest and
fixed by the addition of a visible = true flag in the @JsonTypeInfo
annotation arguments.

Also in this PR is the addition of the iss field to the
SoftwareStatement that represents the software statements created by the
ForgeRock directory. This is necessary as the Postman tests use the
directory server API to obtain a current software statement and that
didn't have an iss field set, although the obtained by the directory UI
did have an iss statement. The lack of the iss field meant that the
SoftwareStatement couldn't be serialised/deserialised and caused
problems for the Postman tests.

Issue: https://github.com/ForgeCloud/ob-deploy/issues/793